requirements.txt
ashfaquecodes==0.5
colorama==0.4.6
mysql-connector-python==8.1.0
protobuf==4.21.12
python-dotenv==1.0.0
